changed the buffer from [u32] to 4*[u8] to seperate colorchannels

+++ src/engine.rs
@@ -405,7 +405,7 @@ impl Engine {
         );
     }
 
-    pub fn run(&mut self, input_handler: &InputHandler) -> &[u32] {
+    pub fn run(&mut self, input_handler: &InputHandler) -> &[u8] {
         //increase frames_count
         self.frame_count += 1;
